backend/pyproject.toml
[tool.poetry]
name = "hospital-device-scan-backend"
version = "0.1.0"
description = "医院设备扫码登记系统后端（FastAPI + PostgreSQL）"
authors = ["ajian <13521048408@126.com>"]
readme = "../readme.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
python-dotenv = "^1.0.0"
alembic = "^1.13.0"
qrcode = "^7.4.0"
jinja2 = "^3.1.0"
httpx = "^0.27.0"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
openpyxl = "^3.1.0"
reportlab = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"